Refactor course dialogs from MaterialCourseList - Place the components into files - Clean up imports
